AttachmentLoader now has create, configure, dispose.
Now the attachment loader should not be disposed until loaded attachments are disposed.

+++ spine-c/include/spine/MeshAttachment.h
@@ -52,7 +52,7 @@ typedef struct spMeshAttachment {
 	float* uvs;
 
 	int trianglesCount;
-	int* triangles;
+	unsigned short* triangles;
 
 	float r, g, b, a;
